Separate multinomial convolution layer into separate file
