+++ layers/__init__.py
@@ -1,0 +1,1 @@
+from layers.multinomial_convolutional import MultinomialConvolutionLayer
